requirements.txt
# django==4.0.2
# django-cors-headers==3.11.0
# djangorestframework==3.13.1
mssql-django==1.4.2
pyodbc==5.1.0
python-dotenv==0.19.2

asgiref==3.7.2
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.6.0